shaky_circles/requirements.txt
pycairo
shapely

pylint
autopep8
